Add CURD operations

+++ user-server/src/main/scala/com/axiom/model/User.scala
@@ -3,7 +3,7 @@ package com.axiom.model
 import zio._
 import zio.json._
 
-case class User(id: Int, firstName: String, lastName: String)
+case class User(id: Long, firstName: String, lastName: String)
 
 object User {
     given JsonEncoder[User] = DeriveJsonEncoder.gen[User]
